refactor(lab): break the step stuff out into it's own module
